provisioned gazelle and renamed mod
