+++ core-be/gqlgen.yml
@@ -35,7 +35,7 @@ resolver:
 # gqlgen will search for any type names in the schema in these go packages
 # if they match it will use them, otherwise it will generate them.
 autobind:
-  - "github.com/buildhub-gg/buildhub/graph/model"
+  - "github.com/buildhub-gg/buildhub/core-be/graph/model"
 
 # This section declares type mapping between the GraphQL and go type systems
 #
